+++ Tests/Get-PuppetNodePCPBrokerDetails.Tests.ps1
@@ -1,0 +1,12 @@
+InModuleScope PSPuppetOrchestrator {
+    Describe 'Get-PuppetPCPNodeBrokerDetails' -Tag unit {
+        function Invoke-RestMethod {}
+        Mock Invoke-RestMethod { return @{name = 'node1'} }
+        context 'returns PCP Broker Details as expected' {
+            it "should return pcp broker details for node1" {
+                (Get-PuppetPCPNodeBrokerDetails -Token 1 -Master 'pm' -Node 'node1').name | should -be 'node1'
+                Assert-MockCalled -CommandName 'Invoke-RestMethod' -Times 1 -Exactly -Scope It
+            }
+        }
+    }
+}