+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.5]
+### Added
+ - added pester tests
+ - token nodes in readme
+ - be consistent with how we write out object with > item
+
 ## [0.1.4]
 ### Added
  - bump build reqs